Small refactor and redefined a new coding approach
Deleted Application and Engine files to simplify the initialization of the engine, I need to focus on getting
the basics working before moving to more advanced concepts.
